소숫점 기능 추가

+++ src/main/java/kr/co/firestock/vo/PortFolioDetail.java
@@ -12,9 +12,9 @@ public class PortFolioDetail {
 
     String portFolioType;
 
-    int portFolioWonMoney;
+    double portFolioWonMoney;
 
-    int portFolioDollarMoney;
+    double portFolioDollarMoney;
 
     String regDt;
